# docs/swagger.yaml
basePath: /
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.GetDiscordAuthURLResponse:
    properties:
      url:
        type: string
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.GetDiscordTokenResponse:
    properties:
      access_token:
        type: string
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.GetUserOutput:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      github_id:
        type: string
      id:
        type: string
      name:
        type: string
      profile:
        type: string
      twitter_id:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API server for the Toybox application.
  title: Toybox API
  version: "1.0"
paths:
  /auth/discord:
    get:
      description: Get Discord authentication URL
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetDiscordAuthURLResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get Discord authentication URL
      tags:
      - auth
  /auth/discord/callback:
    get:
      description: Get Discord token by code
      parameters:
      - description: Discord code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetDiscordTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get Discord token by code
      tags:
      - auth
  /auth/refresh:
    post:
      description: Regenerate token(need refresh token in cookie)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetDiscordTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Regenerate token
      tags:
      - auth
...
  /users/{id}:
    get:
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetUserOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get a user by ID
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"